handle exceptions-backend loggings-cors-swagger-uı

+++ library-backend/src/main/java/com/batuhan/library/service/impl/AddressServiceImpl.java
@@ -2,10 +2,14 @@ package com.batuhan.library.service.impl;
 
 import com.batuhan.library.dto.AddressDTO;
 import com.batuhan.library.entity.Address;
+import com.batuhan.library.entity.Book;
+import com.batuhan.library.exception.ResourceNotFoundException;
 import com.batuhan.library.mapper.AddressMapper;
 import com.batuhan.library.repository.AddressRepository;
 import com.batuhan.library.service.AddressService;
 import lombok.AllArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,12 +19,16 @@ import java.util.stream.Collectors;
 @Service
 @AllArgsConstructor
 public class AddressServiceImpl implements AddressService {
+    private static final Logger logger = LoggerFactory.getLogger(AddressServiceImpl.class);
     private AddressRepository addressRepository;
 
     @Override
     public AddressDTO createAddress(AddressDTO addressDTO) {
+        logger.info("Trying to add a address: {}", addressDTO);
         Address address = AddressMapper.mapToAddressEntity(addressDTO);
+        logger.info("Address entity after the mapping: {}", address);
         address = addressRepository.save(address);
+        logger.info("The address successfully saved in database: {}", address);
         return AddressMapper.mapToAddressDTO(address);
     }
 
@@ -34,15 +42,19 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public AddressDTO getAddressById(Long id) {
-        Optional<Address> optionalAddress = addressRepository.findById(id);
-        Address address = optionalAddress.get();
+        // Optional<Address> optionalAddress = addressRepository.findById(id);
+        // Address address = optionalAddress.get();
+        Address address = addressRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Address", "ID", id));
         return AddressMapper.mapToAddressDTO(address);
     }
 
     @Override
     public AddressDTO updateAddress(AddressDTO addressDTO) {
         Optional<Address> optionalAddress = addressRepository.findById(addressDTO.getId());
-        Address addressToUpdate = optionalAddress.get();
+        Address addressToUpdate = optionalAddress.orElseThrow(
+                () -> new ResourceNotFoundException("Address", "ID", addressDTO.getId())
+        );
         updateAddressEntityFromDTO(addressToUpdate, addressDTO);
         Address updatedAddress = addressRepository.save(addressToUpdate);
         return AddressMapper.mapToAddressDTO(updatedAddress);
@@ -50,6 +62,9 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public void deleteAddress(Long id) {
+        if (!addressRepository.existsById(id)){
+            throw new ResourceNotFoundException("Address", "ID", id);
+        }
         addressRepository.deleteById(id);
     }
